Move connect() to after daemon
If the daemon was started as part of a boot script but before the
networking part of the boot process the loop on connect() before daemon
would hang the entire boot process. This is especially problematic when
dealing with embedded hosts, since they may not have any convenient
reconfiguration methods.